cpu affinity using int64_t
instead of int/32 bit

+++ include/hogl/detail/utils.hpp
@@ -29,6 +29,6 @@ namespace hogl
     
 // return non-zero on error
 // core_id < 0 would not do anything and just return 0
-int setaffinity(pthread_t thread_id, int core_id);
+int setaffinity(pthread_t thread_id, int64_t core_id);
 
 } // hogl
